refactor chain client
